chore: move wallet connect interfaces to common file
