+++ packages/apps/human-app/frontend/src/modules/auth-web3/providers/require-wallet-connect.tsx
@@ -2,8 +2,8 @@ import { useLocation, Navigate } from 'react-router-dom';
 import { createContext } from 'react';
 import { useWalletConnect } from '@/shared/hooks/use-wallet-connect';
 import { routerPaths } from '@/router/router-paths';
-import { type WalletConnectContextConnectedAccount } from '@/shared/types/wallet-connect-context';
 import { PageCardLoader } from '@/shared/components/ui/page-card';
+import { type WalletConnectContextConnectedAccount } from '@/shared/types/wallet-connect';
 
 export const AuthWeb3Context =
   createContext<WalletConnectContextConnectedAccount | null>(null);
